Move out renderData and RenderingData to it's own file
